requirements.txt
asgiref==3.3.1
Django==3.1.7
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-translation-manager==1.1.1
djangorestframework==3.12.4
django-filter==21.1
Markdown==3.3.4
django-parler-rest==2.1